feat: Add inline markdown support to markdown tables and wraptext instead of truncating
